+++ Brovan/Core/Emulation/OS/Windows/Misc/NtCreateMutant.cs
@@ -11,9 +11,9 @@ namespace Brovan.Core.Emulation.OS.Windows
             if (Instance._binary.Architecture == BinaryArchitecture.x64)
             {
                 ulong MutantHandlePtr = Instance.WinHelper.GetArg64(0);
-                ulong DesiredAccess = Instance.WinHelper.GetArg64(1);
+                ulong DesiredAccess = (uint)Instance.WinHelper.GetArg64(1);
                 ulong ObjectAttributesPtr = Instance.WinHelper.GetArg64(2);
-                bool InitialOwner = Instance.WinHelper.GetArg64(3) != 0;
+                bool InitialOwner = (uint)Instance.WinHelper.GetArg64(3) != 0;
 
                 return HandleCreateMutant64(Instance, MutantHandlePtr, DesiredAccess, ObjectAttributesPtr, InitialOwner);
             }
